[player] Fix player hanging when terminate
Very particular path to reproduce the issue :
- Server must stop sending RTSP stream
- Player::stop() and Player::setSources() call after

Reasons:
When source is set, demuxer is "terminated" (method used to reset it). But this method will hang in this particular case, for 10/15 seconds, due to `av_read_frame` (called by auto packet = demuxer.read()` inside `QAVPlayer::doDemux` method)

It seems that FFMPEG doesn't provide a non-blocking equivalent method (flag AVIO_FLAG_NONBLOCK exist but support is really experimental : https://ffmpeg-devel.ffmpeg.narkive.com/k3jOJEmZ/av-read-frame-question-about-blocking).

According to multiple thread on this issue :
- https://stackoverflow.com/questions/14558172/ffmpeg-av-read-frame-need-very-long-time-to-stop
- https://stackoverflow.com/questions/65408224/using-asio-to-implement-interrupt-callback-for-ffmpeg
Using a custom `AVFormatContext::interrupt_callback` is the way to do, good thing is that `QAvDemuxer` already provide it and also already provide an option to perform the `abort`.

So, be sure to enable `abort` property before explicitly ask to wait for the demuxer to finish

(note that setting sources isn't the only way to trigger it, simply destroying player, which call terminate, will trigger this behaviour)